- ERC profiling results - 3D view changes - First version of parasitic - Poly/PolyBase split - Minor DRC changes
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@1614 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/database/geometry/DBMath.java
@@ -42,7 +42,7 @@ public class DBMath extends GenMath {
 	/**
 	 * NDECIMALS number digits to round decimal numbers
 	 */
-	private static double NDECIMALS = 2;
+	private static int NDECIMALS = 2;
 
 	/**
 	 * To return private epsilon used for calculation.
